+++ projects/ngx-enaio/src/lib/interfaces/enaio-document-update-options.ts
@@ -1,0 +1,11 @@
+import { EnaioFieldSchema } from './../enums/enaio-field-schema.enum';
+import { EnaioLocale } from './../enums/enaio-locale.enum';
+
+/**
+ * Update Options
+ */
+export interface EnaioDocumentUpdateOptions {
+  archivable?: boolean;
+  replacefiles?: boolean;
+  objecttypeid?: number;
+}